+++ api/validation.ts
@@ -18,7 +18,6 @@ export const settingsUpdateSchema = z.object({
 export const playerUpdateSchema = z.object({
   name: z.string(),
   date_of_birth: z.string().date(),
-  profile_pic: z.string(),
 });
 
 export const gameUpdateSchema = z.object({