+++ api/services/scoreService.ts
@@ -1,6 +1,5 @@
-import { scores } from "@prisma/client";
 import prisma from "../prismaInstance.js";
-import { getProfilePicUrl } from "./r2.js";
+import { getR2Url } from "./r2.js";
 
 export interface SaveScorePayload {
   points: number;
@@ -64,8 +63,9 @@ export const createProfilePicUrlMapFromHighScores = async (
   const profilePicUrlMap: Record<string, string> = {};
   for (const score of highScores) {
     if (!profilePicUrlMap[score.players?.profile_pic_r2_key!]) {
-      profilePicUrlMap[score.players?.profile_pic_r2_key!] =
-        await getProfilePicUrl(score.players?.profile_pic_r2_key!);
+      profilePicUrlMap[score.players?.profile_pic_r2_key!] = await getR2Url(
+        score.players?.profile_pic_r2_key!
+      );
     }
   }
   return profilePicUrlMap;