+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.1.4 -
 
 - Update Vercel CLI to 46.0.0
-
+- Move profile pictures to Cloudflare R2 Object Storage
+- Rename ``/points`` endpoint to ``/high-scores`` and include profile picture URL
 
 ## 0.1.3 - 12/08/2025 13:10